lambda_go/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-go-alpha==2.80.0a0
